model/django_model/requirements.txt
django
psycopg2
djangorestframework 
django-cors-headers
djangorestframework-simplejwt
pytest
pytest-django